requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
datasets==2.21.0
dill==0.3.8
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.4.1/en_core_web_trf-3.4.1-py3-none-any.whl
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.6.1
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.0
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
packaging==24.2
pandas==1.5.2
pillow==10.2.0
propcache==0.2.0
pyarrow==17.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
spacy-alignments==0.8.6
spacy-transformers==1.1.8
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
torchvision==0.19.1+cu121
tqdm==4.67.1
transformers==4.44.2
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
xxhash==3.5.0
yarl==1.15.2
zstandard==0.19.0
